requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
Authlib==1.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.15.1
fonttools==4.53.0
fsspec==2024.6.0
grpcio==1.63.0
grpcio-health-checking==1.63.0
grpcio-tools==1.63.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kneed==0.8.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.26.0
openpyxl==3.1.2
orjson==3.10.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==24.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.4.28
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scikit-learn-extra==0.3.0
scikit-optimize==0.10.2
scipy==1.13.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
thinc==8.2.4
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
validators==0.28.1
wasabi==1.1.3
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.4.1
weaviate-client==4.5.7
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
XlsxWriter==3.2.0
